Delete: Eliminacion del archivo user_proyects_model.py. Dicha funcionalidad fue agregada en user_db_model.py New: Creacion del archivo de test para su posterior implementacion
